Parameter Make.2-Parse

val parse_logic : Dolmen.Std.Statement.t list -> State.t -> Dolmen_loop.Logic.language Dolmen_loop.State.file -> State.t * ( State.t -> State.t * Dolmen.Std.Statement.t option )

Parsing function. Reads a list of prelude statements, and the state and returns a tuple of the new state (including the detected input language), together with a statement generator.

val parse_response : Dolmen.Std.Answer.t list -> State.t -> Dolmen_loop.Response.language Dolmen_loop.State.file -> State.t * ( State.t -> State.t * Dolmen.Std.Answer.t option )

Parsing function. Reads a list of prelude statements, and the state and returns a tuple of the new state (including the detected input language), together with a statement generator.

val expand : State.t -> Dolmen.Std.Statement.t -> State.t * [ `Ok | `Gen of ( State.t -> State.t -> State.t ) * ( State.t -> State.t * Dolmen.Std.Statement.t option ) ]

Expand statements (such as includes). Returns the new state, and either:

`Ok , which means the statement can be propagated as is

`Gen (flat, g) , if the statement expands into a generator g. The bool flat indicates wether the statements in g should be treated as a single group of statements (with regards to timeouts, etc...), or as a list of independant statements (each with its own timeout...).